requirements.txt
Django==4.2.27
pillow==10.4.0